Move shared help text out of the CLI entry point
The server-details JSON help builders (and the schema base URL they link to)
now live in src/cli/help-text.ts, so index.ts holds command wiring rather
than help-text construction.

Refs #340
